Fix possible non-determinism
I was treating the rb field as equivalent to the genomic pos, which is
only true for forward reads.  It is possible this led to some
unexplained indeterminism.  To avoid this, I now store the score, the
rb and the qb of the best alignment, hoping this is enough to sort with.